# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "kmf tech task currency",
        "contact": {}
    },
    "paths": {
        "/currency/save/{date}": {
            "get": {
                "description": "Метод для сохранение записей из НацБанка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SaveRates"
                ],
                "summary": "Сохранить записи",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SuccessResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResp"
                        }
                    }
                }
            }
        },
        "/currency/save/{date}/{code}": {
            "get": {
                "description": "Метод для получение записей из базы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetRates"
                ],
                "summary": "Получить записи",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Currency"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.SuccessResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "entity.Currency": {
            "type": "object",
            "properties": {
                "adate": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}